Refactor viewport overlays and pathtracer maps
Introduce overlay rendering and multi-map pathtracer support. Key changes:

- Add Registry.pinned to keep nodes visible regardless of selection. Update reset() to clear pinned.
- MeshProvider gains default isOverlay(); several nodes (Pose, Camera, DHParameter, HingeJoint, LinearJoint, MeshInstance) implement overlay behavior and update isActive() to honor selection or pinned lists.
- DrawMeshes refactor: add overlayShader and drawOverlays() to render diagnostic/overlay meshes without depth testing, skip overlays during shadow generation, and avoid casting shadows for non-shadowing/overlay providers. Outline selection now operates on a filtered selected list returned from keepOnlySelectedMeshMaterials().
- Remove obsolete render passes: DrawDepthBuffer, DrawPoses, DrawStencilBuffer.
- Improve DrawBoundingBoxes: compute center/scale from AABB and build a model matrix for a unit-cube mesh.
- OpenGL shader resource path updates (OpenGL3Panel, DrawBackground, DrawBoundingBoxes) and other minor GL cleanups.
- PathTracer: replace single image with multiple maps (Color, Depth, Normal, Sample) stored in a Map; expose getMaps(); create and populate sampleMap; adjust worker to use colorMap; fix sample accumulation logic in RayXY (cast to double).
- PathTracerPanel: use combo box populated from pathTracer.getMaps(), fix save logic to use selected map, move runtime label to toolbar.
- RO3Frame: reorganize default docking order to add BrainView, Donatello, PathTracerPanel, etc.
- Update Changelog with Unreleased section and BabaIsYou app entry.

These changes separate overlay/diagnostic visuals from normal scene geometry, centralize multiple render targets for the pathtracer, and simplify shadow/depth handling.

+++ src/main/java/com/marginallyclever/ro3/apps/pathtracer/RayXY.java
@@ -52,7 +52,7 @@ public class RayXY {
         samples++;
         // recalculate the average.
         radianceAverage.set(radianceSum);
-        radianceAverage.scale(1.0/samples);
+        radianceAverage.scale(1.0/(double)samples);
         // apply exposure
         exposureMap(radianceAverage,exposure);
         // tone map the result.
